third_party/WebKit/LayoutTests/svg/css/svg-paint-order-expected.txt

Test paint-order.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Test pre-normalized correct variants of 'paint-order'
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill markers stroke"
PASS getComputedStyle(text).paintOrder is "stroke fill markers"
PASS getComputedStyle(text).paintOrder is "stroke markers fill"
PASS getComputedStyle(text).paintOrder is "markers stroke fill"
PASS getComputedStyle(text).paintOrder is "markers fill stroke"

Test correct single keyword value of 'paint-order'
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "stroke fill markers"
PASS getComputedStyle(text).paintOrder is "markers fill stroke"

Test correct dual keyword values of 'paint-order'
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill markers stroke"
PASS getComputedStyle(text).paintOrder is "stroke fill markers"
PASS getComputedStyle(text).paintOrder is "stroke markers fill"
PASS getComputedStyle(text).paintOrder is "markers fill stroke"
PASS getComputedStyle(text).paintOrder is "markers stroke fill"

Test invalid values of 'paint-order'
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill stroke markers"

Test pre-normalized correct variants of 'paint-order' (presentation attribute)
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill markers stroke"
PASS getComputedStyle(text).paintOrder is "stroke fill markers"
PASS getComputedStyle(text).paintOrder is "stroke markers fill"
PASS getComputedStyle(text).paintOrder is "markers stroke fill"
PASS getComputedStyle(text).paintOrder is "markers fill stroke"

Test correct single keyword value of 'paint-order' (presentation attribute)
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "stroke fill markers"
PASS getComputedStyle(text).paintOrder is "markers fill stroke"

Test correct dual keyword values of 'paint-order' (presentation attribute)
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill markers stroke"
PASS getComputedStyle(text).paintOrder is "stroke fill markers"
PASS getComputedStyle(text).paintOrder is "stroke markers fill"
PASS getComputedStyle(text).paintOrder is "markers fill stroke"
PASS getComputedStyle(text).paintOrder is "markers stroke fill"

Test invalid values of 'paint-order' (presentation attribute)
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS getComputedStyle(text).paintOrder is "fill stroke markers"
PASS successfullyParsed is true

TEST COMPLETE
